Moved NLLS-specific methods from NonlinearOptimizer class to subclass.
